operators type update compatibility fixes in app
