+++ app/packages/operators/src/index.ts
@@ -1,4 +1,5 @@
 export { default as OperatorBrowser } from "./OperatorBrowser";
 export { default as OperatorPrompt } from "./OperatorPrompt";
-export { registerOperator, loadOperators } from "./operators";
+export { registerOperator } from "./operators";
 export { useOperatorExecutor } from "./state";
+export { loadOperators } from "./built-in-operators";
