+++ app/packages/operators/src/operators.ts
@@ -1,16 +1,6 @@
 import { getFetchFunction, ServerError } from "@fiftyone/utilities";
+import { CallbackInterface } from "recoil";
 import * as types from "./types";
-import { CallbackInterface, useRecoilValue } from "recoil";
-import * as fos from "@fiftyone/state";
-import {
-  SpaceNode,
-  usePanels,
-  useSpaces,
-  useSpaceNodes,
-} from "@fiftyone/spaces";
-
-import copyToClipboard from "copy-to-clipboard";
-import { useOperatorExecutor } from ".";
 
 export class ExecutionContext {
   public state: CallbackInterface;
@@ -57,44 +47,47 @@ class OperatorResult {
   }
 }
 
-class OperatorDefinition {
-  constructor(public description: string) {
-    this.inputs = new types.ObjectType();
-    this.outputs = new types.ObjectType();
-  }
-  public inputs: types.ObjectType;
-  public outputs: types.ObjectType;
-  addInputProperty(property: types.Property) {
-    this.inputs.addProperty(property);
-  }
-  addOutputProperty(property: types.Property) {
-    this.outputs.addProperty(property);
-  }
-  static fromJSON(json: any) {
-    const def = new OperatorDefinition(json.description);
-    def.inputs = types.ObjectType.fromJSON(json.inputs);
-    def.outputs = types.ObjectType.fromJSON(json.outputs);
-    return def;
-  }
-}
-
 export class Operator {
-  public definition: OperatorDefinition;
-  constructor(public name: string, description: string) {
-    this.name = name;
-    this.definition = new OperatorDefinition(description);
+  public definition: types.ObjectType;
+  constructor(
+    public name: string,
+    public description: string,
+    public inputView?,
+    public outputView?
+  ) {
+    this.definition = new types.ObjectType();
+    this.definition.defineProperty("inputs", new types.ObjectType(), {
+      view: inputView,
+    });
+    this.definition.defineProperty("outputs", new types.ObjectType(), {
+      view: inputView,
+    });
+  }
+  get inputs(): types.Property {
+    return this.definition.getProperty("inputs");
+  }
+  get outputs(): types.Property {
+    return this.definition.getProperty("outputs");
+  }
+  defineInputProperty(name, type, options?) {
+    const inputsType = this.inputs.type as types.ObjectType;
+    return inputsType.defineProperty(name, type, options);
+  }
+  defineOutputProperty(name, type, options) {
+    const outputsType = this.outputs.type as types.ObjectType;
+    return outputsType.defineProperty(name, type, options);
   }
   needsUserInput() {
-    return this.definition.inputs.properties.length > 0;
+    const inputsType = this.inputs.type as types.ObjectType;
+    return inputsType.properties.size > 0;
   }
   needsResolution() {
-    return this.definition.inputs.needsResolution();
+    const inputsType = this.inputs.type as types.ObjectType;
+    return inputsType.needsResolution();
   }
   needsOutput(result: OperatorResult) {
-    if (
-      this.definition.outputs.properties.length > 0 &&
-      result.hasOutputContent()
-    ) {
+    const outputType = this.outputs.type as types.ObjectType;
+    if (outputType.properties.size > 0 && result.hasOutputContent()) {
       return true;
     }
     if (result.error) {
@@ -106,28 +99,28 @@ export class Operator {
     // This can be overridden to use hooks in the execute function
     return {};
   }
-  async resolveInputRemove(ctx: ExecutionContext) {
-    const { inputs } = this.definition;
+  async resolveInputRemote(ctx: ExecutionContext) {
     return resolveRemoteType(this.name, ctx, "inputs");
   }
   async resolveInput(ctx: ExecutionContext) {
-    const { inputs } = this.definition;
-    if (inputs.needsResolution()) {
+    const inputsType = this.inputs.type as types.ObjectType;
+    if (inputsType.needsResolution()) {
       if (this.isRemote) {
-        return this.resolveInputRemove(ctx);
+        return this.resolveInputRemote(ctx);
       }
+
       const resolvedInputs = new types.ObjectType();
-      for (const property of inputs.properties) {
+      inputsType.properties.forEach(async (property, name) => {
         if (property.hasResolver) {
           const resolved = await property.resolver(property, ctx);
-          resolvedInputs.addProperty(resolved);
+          resolvedInputs.addProperty(name, resolved);
         } else {
-          resolvedInputs.addProperty(property);
+          resolvedInputs.addProperty(name, property);
         }
-      }
-      return resolvedInputs;
+      });
+      return new types.Property(resolvedInputs, { view: this.inputView });
     }
-    return inputs;
+    return this.inputs;
   }
   async execute(ctx: ExecutionContext) {
     throw new Error(`Operator ${this.name} does not implement execute`);
@@ -139,8 +132,13 @@ export class Operator {
     return operator;
   }
   static fromJSON(json: any) {
+    const { inputs, outputs } = json.definition.properties;
     const operator = new Operator(json.name, json.description);
-    operator.definition = OperatorDefinition.fromJSON(json.definition);
+    operator.definition.addProperty("inputs", types.Property.fromJSON(inputs));
+    operator.definition.addProperty(
+      "outputs",
+      types.Property.fromJSON(outputs)
+    );
     return operator;
   }
 }
@@ -291,452 +289,5 @@ export async function resolveRemoteType(
     throw new Error(typeAsJSON.error);
   }
 
-  return types.typeFromJSON(typeAsJSON);
-}
-
-//
-// BUILT-IN OPERATORS
-//
-class ReloadSamples extends Operator {
-  constructor() {
-    super("reload_samples", "Reload samples from the dataset");
-  }
-  async execute({ state }: ExecutionContext) {
-    const refresherTick = await state.snapshot.getPromise(fos.refresher);
-    state.set(fos.refresher, refresherTick + 1);
-  }
-}
-
-class ClearSelectedSamples extends Operator {
-  constructor() {
-    super("clear_selected_samples", "Clear selected samples");
-  }
-  useHooks(): {} {
-    return {
-      setSelected: fos.useSetSelected(),
-    };
-  }
-  async execute({ hooks, state }: ExecutionContext) {
-    // needs to mutate the server / session
-    hooks.setSelected([]);
-    state.reset(fos.selectedSamples);
-  }
-}
-
-class CopyViewAsJSON extends Operator {
-  constructor() {
-    super("copy_view_as_json", "Copy view as JSON");
-  }
-  async execute({ state }: ExecutionContext) {
-    const view = await state.snapshot.getPromise(fos.view);
-    const json = JSON.stringify(view, null, 2);
-    copyToClipboard(json);
-  }
-}
-
-class ViewFromJSON extends Operator {
-  constructor() {
-    super("view_from_json", "Create view from clipboard");
-  }
-  async execute({ state }: ExecutionContext) {
-    const text = await navigator.clipboard.readText();
-    try {
-      const view = JSON.parse(text);
-      state.set(fos.view, view);
-    } catch (e) {
-      console.error("Error parsing JSON", e);
-    }
-  }
-}
-
-class OpenPanel extends Operator {
-  constructor() {
-    super("open_panel", "Open a panel");
-    // todo: dynamically generate the list
-    this.definition.inputs.addProperty(
-      new types.Property(
-        "name",
-        new types.Enum(["Histograms", "Embeddings"]),
-        "Name of the panel",
-        true,
-        ""
-      )
-    );
-    this.definition.inputs.addProperty(
-      new types.Property(
-        "isActive",
-        new types.Boolean(),
-        "Auto-select on open",
-        true,
-        false
-      )
-    );
-  }
-  useHooks(): object {
-    const { FIFTYONE_SPACE_ID } = fos.constants;
-    const availablePanels = usePanels();
-    const { spaces } = useSpaces(FIFTYONE_SPACE_ID);
-    const openedPanels = useSpaceNodes(FIFTYONE_SPACE_ID);
-    return { availablePanels, openedPanels, spaces };
-  }
-  findFirstPanelContainer(node: SpaceNode): SpaceNode {
-    if (node.isPanelContainer()) {
-      return node;
-    }
-
-    if (node.hasChildren()) {
-      return this.findFirstPanelContainer(node.firstChild());
-    }
-  }
-  async execute({ hooks, params }: ExecutionContext) {
-    console.log(hooks, params);
-    const { spaces } = hooks;
-    const { name, isActive } = params;
-    const targetSpace = this.findFirstPanelContainer(spaces.root);
-    if (!targetSpace) {
-      return console.error("No panel container found");
-    }
-    const newNode = new SpaceNode();
-    newNode.type = name;
-    // add panel to the default space as an inactive panels
-    spaces.addNodeAfter(targetSpace, newNode, isActive);
-  }
-}
-
-class OpenAllPanels extends Operator {
-  constructor() {
-    super("open_all_panel", "Open all panels");
-  }
-  useHooks(): object {
-    const { FIFTYONE_SPACE_ID } = fos.constants;
-    const availablePanels = usePanels();
-    const openedPanels = useSpaceNodes(FIFTYONE_SPACE_ID);
-    const openPanelOperator = useOperatorExecutor("open_panel");
-    return { availablePanels, openedPanels, openPanelOperator };
-  }
-  async execute({ hooks }: ExecutionContext) {
-    const { availablePanels, openedPanels, openPanelOperator } = hooks;
-    const openedPanelsTypes = openedPanels.map(({ type }) => type);
-    for (const panel of availablePanels) {
-      const { name } = panel;
-      if (openedPanelsTypes.includes(name)) continue;
-      openPanelOperator.execute({ name, isActive: false });
-    }
-  }
-}
-
-// class FindSpace extends Operator {
-//   constructor() {
-//     super("find_space", "Find space");
-//   }
-//   findFirstPanelContainer(node: SpaceNode): SpaceNode {
-//     if (node.isPanelContainer()) {
-//       return node;
-//     }
-//     if (node.hasChildren()) {
-//       return this.findFirstPanelContainer(node.firstChild());
-//     }
-//   }
-//   useHooks(): object {
-//     const { FIFTYONE_SPACE_ID } = fos.constants;
-//     const { spaces } = useSpaces(FIFTYONE_SPACE_ID);
-//     return { spaces };
-//   }
-//   async execute({ hooks }: ExecutionContext) {
-//     console.log("Find space invoked");
-//     const { spaces } = hooks;
-//     return this.findFirstPanelContainer(spaces.root);
-//   }
-// }
-
-class ClosePanel extends Operator {
-  constructor() {
-    super("close_panel", "Close a panel");
-    // todo: dynamically generate the list
-    this.definition.inputs.addProperty(
-      new types.Property(
-        "name",
-        new types.Enum(["Histograms", "Embeddings"]),
-        "Name of the panel",
-        true,
-        ""
-      )
-    );
-  }
-  useHooks(): object {
-    const { FIFTYONE_SPACE_ID } = fos.constants;
-    const { spaces } = useSpaces(FIFTYONE_SPACE_ID);
-    const openedPanels = useSpaceNodes(FIFTYONE_SPACE_ID);
-    return { openedPanels, spaces };
-  }
-  async execute({ hooks, params }: ExecutionContext) {
-    const { openedPanels, spaces } = hooks;
-    const { name, id } = params;
-    const targetPanel = openedPanels.find(
-      (panel) => id === panel.id || name === panel.type
-    );
-    if (!targetPanel)
-      return console.error(
-        `Opened panel with ${id ? "id" : "name"} "${
-          id || name
-        }" cannot be found`
-      );
-    spaces.removeNode(targetPanel);
-  }
-}
-
-class CloseAllPanels extends Operator {
-  constructor() {
-    super("close_all_panel", "Close all panels");
-  }
-  useHooks(): object {
-    const { FIFTYONE_SPACE_ID } = fos.constants;
-    const openedPanels = useSpaceNodes(FIFTYONE_SPACE_ID);
-    const closePanel = useOperatorExecutor("close_panel");
-    return { openedPanels, closePanel };
-  }
-  async execute({ hooks }: ExecutionContext) {
-    const { openedPanels, closePanel } = hooks;
-    for (const panel of openedPanels) {
-      // do not close pinned, root or space panel
-      if (panel.pinned || panel.isRoot() || panel.isSpace()) continue;
-      closePanel.execute(panel);
-    }
-  }
-}
-
-class OpenDataset extends Operator {
-  constructor() {
-    super("open_dataset", "Open Dataset");
-    const datasetProprety = this.definition.inputs.addProperty(
-      new types.Property(
-        "dataset",
-        new types.Enum([]),
-        "Name of the dataset",
-        true
-      )
-    );
-    datasetProprety.resolver = (property: any, ctx: ExecutionContext) => {
-      console.log("datasetProprety.resolver", ctx);
-      property.type = new types.Enum(ctx.hooks.availableDatasets);
-      return property;
-    };
-  }
-  useHooks(ctx: ExecutionContext): object {
-    // const useSearch = getUseSearch();
-
-    // const {values, total} = useSearch(ctx.params.dataset);
-    return {
-      availableDatasets: ["quickstart", "quickstart-geo", "quickstart-groups"],
-      setDataset: fos.useSetDataset(),
-    };
-  }
-  async execute({ hooks, params }: ExecutionContext) {
-    hooks.setDataset(params.dataset);
-  }
-}
-
-class ClearView extends Operator {
-  constructor() {
-    super("clear_view", "Clear view bar");
-  }
-  async execute({ state }: ExecutionContext) {
-    state.reset(fos.view);
-  }
-}
-class ClearSidebarFilters extends Operator {
-  constructor() {
-    super("clear_sidebar_filters", "Clear sidebar filters");
-  }
-  async execute({ state }: ExecutionContext) {
-    state.reset(fos.filters);
-  }
-}
-
-class ClearAllStages extends Operator {
-  constructor() {
-    super("clear_all_stages", "Clear all selections, filters, and view");
-  }
-  useHooks(): {} {
-    return {
-      setSelected: fos.useSetSelected(),
-      resetExtended: fos.useResetExtendedSelection(),
-    };
-  }
-  async execute({ state, hooks }: ExecutionContext) {
-    state.reset(fos.view);
-    state.reset(fos.filters);
-    hooks.resetExtended();
-    state.reset(fos.selectedSamples);
-    hooks.setSelected([]);
-  }
-}
-
-class RefreshColors extends Operator {
-  constructor() {
-    super("refresh_colors", "Refresh colors");
-  }
-  async execute({ state }: ExecutionContext) {
-    const modal = await state.snapshot.getPromise(fos.modal);
-    console.log("modal", modal);
-    const colorsSeed = await state.snapshot.getPromise(fos.colorSeed(!!modal));
-    state.set(fos.colorSeed(!!modal), colorsSeed + 1);
-  }
-}
-
-class ShowSelectedSamples extends Operator {
-  constructor() {
-    super("show_selected_samples", "Show selected samples");
-  }
-  useHooks(): {} {
-    return {
-      setSelected: fos.useSetSelected(),
-    };
-  }
-  async execute({ hooks, state }: ExecutionContext) {
-    const selectedSamples = await state.snapshot.getPromise(
-      fos.selectedSamples
-    );
-    console.log("selectedSamples", selectedSamples);
-    state.set(fos.extendedSelection, {
-      selection: Array.from(selectedSamples),
-      scope: "global",
-    });
-  }
-}
-
-class ConvertExtendedSelectionToSelectedSamples extends Operator {
-  constructor() {
-    super(
-      "convert_extended_selection_to_selected_samples",
-      "Convert extended selection to selected samples"
-    );
-  }
-  useHooks(): {} {
-    return {
-      setSelected: fos.useSetSelected(),
-      resetExtended: fos.useResetExtendedSelection(),
-    };
-  }
-  async execute({ hooks, state }: ExecutionContext) {
-    const extendedSelection = await state.snapshot.getPromise(
-      fos.extendedSelection
-    );
-    console.log("extendedSelection", extendedSelection);
-    state.set(fos.selectedSamples, new Set(extendedSelection.selection));
-    state.set(fos.extendedSelection, { selection: null });
-    hooks.setSelected(extendedSelection.selection);
-    hooks.resetExtended();
-  }
-}
-
-// an operator that sets selected samples based on ctx.params
-class SetSelectedSamples extends Operator {
-  constructor() {
-    super("set_selected_samples", "Set selected samples");
-    this.definition.inputs.addProperty(
-      new types.Property(
-        "samples",
-        new types.List(new types.SampleID()),
-        "Samples",
-        true
-      )
-    );
-  }
-  useHooks(): {} {
-    return {
-      setSelected: fos.useSetSelected(),
-    };
-  }
-  async execute({ hooks, params }: ExecutionContext) {
-    hooks.setSelected(params.samples);
-  }
-}
-
-class DynamicFormExample extends Operator {
-  constructor() {
-    super("dynamic_form_example", "Dynamic Form Example");
-    this.definition.inputs.addProperty(
-      new types.Property(
-        "mode",
-        new types.Enum(["simple", "advanced"]),
-        "Mode",
-        true,
-        "simple"
-      )
-    );
-  }
-  useHooks(ctx: ExecutionContext): {} {
-    const [sampleId] = useRecoilValue(fos.selectedSamples) || [];
-    return {
-      sampleId,
-    };
-  }
-  async resolveInput(ctx: ExecutionContext) {
-    const inputs = new types.ObjectType();
-    inputs.addProperty(
-      new types.Property(
-        "mode",
-        new types.Enum(["simple", "advanced"]),
-        "Mode",
-        true,
-        "simple"
-      )
-    );
-    inputs.addProperty(
-      new types.Property("name", new types.String(), "Name", true)
-    );
-    if (ctx.params.mode === "advanced") {
-      inputs.addProperty(
-        new types.Property(
-          "sampleId",
-          new types.String(),
-          "Sample ID",
-          true,
-          ctx.hooks.sampleId
-        )
-      );
-      inputs.addProperty(
-        new types.Property("threshold", new types.Number(), "Threshold", true)
-      );
-      if (ctx.params.threshold > 5) {
-        inputs.addProperty(
-          new types.Property("clamp", new types.Boolean(), "Clamping", true)
-        );
-      }
-    }
-    return inputs;
-  }
-}
-
-export function registerBuiltInOperators() {
-  try {
-    registerOperator(new CopyViewAsJSON());
-    registerOperator(new ViewFromJSON());
-    registerOperator(new ReloadSamples());
-    registerOperator(new ClearSelectedSamples());
-    registerOperator(new OpenAllPanels());
-    registerOperator(new CloseAllPanels());
-    registerOperator(new OpenDataset());
-    registerOperator(new ClearView());
-    registerOperator(new ClearSidebarFilters());
-    registerOperator(new ClearAllStages());
-    registerOperator(new RefreshColors());
-    registerOperator(new ShowSelectedSamples());
-    registerOperator(new ConvertExtendedSelectionToSelectedSamples());
-    registerOperator(new DynamicFormExample());
-    registerOperator(new OpenPanel());
-    registerOperator(new OpenAllPanels());
-    registerOperator(new ClosePanel());
-    registerOperator(new CloseAllPanels());
-    // registerOperator(new FindSpace());
-  } catch (e) {
-    console.error("Error registering built-in operators");
-    console.error(e);
-  }
-}
-
-export async function loadOperators() {
-  registerBuiltInOperators();
-  await loadOperatorsFromServer();
+  return types.Property.fromJSON(typeAsJSON);
 }
